# AppMobilenBlog/AppMobilenBlog/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RestAPInBlog",
    "version": "1.0"
  },
  "paths": {
    "/api/Comment": {
      "get": {
        "tags": [
          "Comment"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentForView"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentForView"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentForView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Comment"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CommentForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForView"
                }
              }
            }
          }
        }
      }
    },
    "/api/Comment/{id}": {
      "get": {
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CommentForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CommentForView": {
        "type": "object",
        "properties": {
          "commentId": {
            "type": "integer",
            "format": "int32"
          },
          "postId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "userId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "commentDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}